syscall: drop readonly annotation from syscall0()
The syscall0() backends currently annotate the inline-asm as `readonly`,
presumably because without argument there is little reason to assume the
system call writes to memory. This patch drops this annotation under the
assumption that there are scenarios where those calls will write to
memory.

While we have not really found an obvious violation of this, there are
definitely system calls that do affect mapped memory. For instance, the
VDSO is updated on a wide range of system calls, fork(2) can write
process identifiers to memory, sync(2) can cause mapped ring-buffers to
clear, etc.

While it would probably be safe to argue that all those scenarios work
with volatile memory and should thus be accessed with care, anyway, it
is a fragile assumption to rely on. And we prefer being correct than
slightly faster, so drop this annotation. If a particular system call
does benefit from this, we can always add alternative entry-points.
However, it is likely that all VDSO-optimizations would outperform any
syscall-entry, so maybe the better alternative is to map the VDSO.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>