Add arm thumb-mode support
r7 is used as the stack pointer in thumb-mode, so we must save/restore
that register when doing the syscall.

Rust support for checking the thumb-mode target feature is severely
lacking on stable, so we must use a build.rs script to automatically
determine if we can even compile for thumb-mode.

`#[instruction_set(arm::*)]` is also useless because it only supports
ARMv4 and ARMv5te, and isn't really conditional compilation. We'd still
need to detect if we're in thumb-mode at runtime.